+++ .editorconfig
@@ -18,3 +18,4 @@ indent_size = 2
 end_of_line = lf
 trim_trailing_whitespace = true
 insert_final_newline = true
+
