+++ js/constans.js
@@ -1,0 +1,35 @@
+const MAX_PHOTOS_NUMBER = 25;
+const DESCRIPTION = [
+  'Вид на отель',
+  'Указатель пляжа',
+  'Вид на скалы и море',
+  'Девушка в купальнике',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = ['Александр', 'Павел', 'Сергей', 'Владимир'];
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const MAX_COMMENTS = 20;
+const MAX_COMMENTS_TOTAL = 5000;
+const MAX_AVATAR_NUMBER = 6;
+
+export {
+  MAX_PHOTOS_NUMBER,
+  DESCRIPTION,
+  MESSAGES,
+  NAMES,
+  MIN_LIKES,
+  MAX_LIKES,
+  MAX_COMMENTS,
+  MAX_COMMENTS_TOTAL,
+  MAX_AVATAR_NUMBER,
+};
