+++ js/functions.js
@@ -1,44 +1,43 @@
-/* eslint-disable quotes */
 //1
 function checkLengthString(word, number) {
   return word.length > number;
 }
 
-checkLengthString("airport", 5);
-checkLengthString("cat", 4);
+console.log(checkLengthString('airport', 5));
+checkLengthString('cat', 4);
 
 //2
 function getPalindrome(string) {
   const reverse = reverseString(string);
   return string === reverse;
 }
-getPalindrome("коm");
+getPalindrome('коm');
 
 function reverseString(str) {
-  return str.split("").reverse().join("");
+  return str.split('').reverse().join('');
 }
 
-reverseString("hello");
+reverseString('hello');
 
 //3
 function getNumber(str) {
-  const newArray = str.split("");
-  let newString = "";
+  const newArray = str.split('');
+  let newString = '';
   newArray.forEach((letter) => {
-    newString += Number.isInteger(letter * 1) ? letter : "";
+    newString += Number.isInteger(letter * 1) ? letter : '';
   });
   return newString.length > 0 ? newString : NaN;
 }
-getNumber("rtrtrtrt");
+getNumber('rtrtrtrt');
 
 //4
 function fillString(str, minLength, addSymbol) {
   let i = 0;
-  let strSymbol = "";
+  let strSymbol = '';
   while (str.length + strSymbol.length <= minLength) {
     strSymbol += addSymbol[i];
     i = i + 1 < addSymbol.length ? i + 1 : 0;
   }
   return strSymbol + str;
 }
-fillString("qwerty", 4, "0");
+fillString('qwerty', 4, '0');
